feat: implement browserViews contribute code snippet
