+++ packages/core-browser/src/extensions/index.ts
@@ -1,1 +1,3 @@
 export * from './extensions-point.service';
+export * from './schema/vscodeExtensionPackageSchema';
+export * from './schema/opensumiExtensionPackageSchema';
