+++ packages/extension/src/browser/sumi/contributes/browser-views.ts
@@ -1,5 +1,6 @@
 import { Injectable, Autowired } from '@opensumi/di';
 import { getIcon } from '@opensumi/ide-core-browser';
+import { browserViews } from '@opensumi/ide-core-browser/lib/extensions/schema/browserViews';
 import { DisposableCollection, getDebugLogger } from '@opensumi/ide-core-common';
 import { IMainLayoutService } from '@opensumi/ide-main-layout';
 import { IIconService } from '@opensumi/ide-theme';
@@ -42,6 +43,8 @@ export class BrowserViewContributionPoint extends VSCodeContributePoint<KtViewsC
 
   @Autowired(IIconService)
   iconService: IIconService;
+
+  static schema = browserViews.schema;
 
   private disposableCollection: DisposableCollection = new DisposableCollection();
