+++ packages/i18n/src/common/contributes/zh-CN.lang.ts
@@ -1,0 +1,13 @@
+export const browserViews = {
+  "kaitianContributes.browserViews": "向编辑器提供前端视图",
+  "kaitianContributes.browserViews.left": "在左侧活动栏容器提供视图",
+  "kaitianContributes.browserViews.right": "在右侧活动栏容器提供视图",
+  "kaitianContributes.browserViews.bottom": "在底部标签栏容器提供视图（不会渲染图标，所以需要提供标题）",
+  "kaitianContributes.browserViews.location.custom": "向 \"{0}\" 容器提供视图",
+  "kaitianContributes.browserViews.view.id": "视图 id, browser/index 中需要导出一个与此 id 相同名称的组件",
+  "kaitianContributes.browserViews.view.icon": "图标名称，参考: [内置图标集](https://docs.antfin-inc.com/iconfont-demo/)",
+  "kaitianContributes.browserViews.view.iconPath": "本地图标相对路径",
+  "kaitianContributes.browserViews.view.title": "视图标题（当 type 为 bottom 时，将展示）",
+  "kaitianContributes.browserViews.view.titleComponentId": "自定义视图标题组件 id, 特性与普通的 browserView 完全一致",
+  "kaitianContributes.browserViews.view.expanded": "是否需要全屏展开，这会遮盖掉 editor 区域",
+}
