+++ packages/i18n/src/common/zh-CN.lang.ts
@@ -1,3 +1,5 @@
+import { browserViews } from "./contributes/zh-CN.lang";
+
 export const localizationBundle = {
   languageId: 'zh-CN',
   languageName: 'Chinese',
@@ -1001,5 +1003,8 @@ export const localizationBundle = {
     'connection.stop.rtt': '开发人员工具：关闭通信延迟检查',
 
     'debug.terminal.label': '创建 Javascript Debug Terminal',
+
+    // extension contribute
+    ...browserViews
   },
 };
